+++ src/agents/agentHooks.go
@@ -3,9 +3,9 @@ package agents
 import (
 	"fmt"
 
-	agentforge "github.com/thinktwice/agentForge/src"
-	"github.com/thinktwice/agentForge/src/core"
-	"github.com/thinktwice/agentForge/src/llms"
+	agentforge "github.com/thinktwiceco/agent-forge/src"
+	"github.com/thinktwiceco/agent-forge/src/core"
+	"github.com/thinktwiceco/agent-forge/src/llms"
 )
 
 // type OnContextBuildHook func(agentContext *core.AgentContext) error